requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
absl-py=0.15.0=pyhd3eb1b0_0
aiohttp=3.8.1=py38h7f8727e_0
aiosignal=1.2.0=pyhd3eb1b0_0
antlr4-python3-runtime=4.8=pypi_0
argon2-cffi=20.1.0=py38h27cfd23_1
async-timeout=4.0.1=pyhd3eb1b0_0
async_generator=1.10=pyhd3eb1b0_0
attrs=21.2.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
black=21.11b1=pypi_0
blas=1.0=mkl
bleach=4.0.0=pyhd3eb1b0_0
blinker=1.4=py38h06a4308_0
brotlipy=0.7.0=py38h27cfd23_1003
c-ares=1.18.1=h7f8727e_0
ca-certificates=2021.10.26=h06a4308_2
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2021.10.8=py38h06a4308_2
cffi=1.14.6=py38h400218f_0
charset-normalizer=2.0.7=pypi_0
click=8.0.3=pypi_0
colorama=0.4.4=pyh9f0ad1d_0
cryptography=3.4.8=py38hd23ed53_0
cudatoolkit=11.0.221=h6bb024c_0
cycler=0.11.0=pypi_0
dataclasses=0.8=pyh6d0b6a4_7
dbus=1.13.18=hb2f20db_0
debugpy=1.5.1=py38h295c915_0
decorator=5.1.0=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.3=py38_0
expat=2.4.1=h2531618_2
ffmpeg=1.4=pypi_0
flake8=4.0.1=pypi_0
flake8-bugbear=21.9.2=pypi_0
flake8-comprehensions=3.7.0=pypi_0
fontconfig=2.13.1=h6c09931_0
fonttools=4.28.2=pypi_0
freetype=2.11.0=h70c0345_0
frozenlist=1.2.0=py38h7f8727e_0
fvcore=0.1.5.post20210804=pyhd8ed1ab_0
giflib=5.2.1=h7b6447c_0
glib=2.69.1=h5202010_0
google-auth=1.33.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.1=py_2
grpcio=1.42.0=py38hce63b2e_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
hydra-core=1.1.1=pypi_0
icu=58.2=he6710b0_3
idna=3.3=pyhd3eb1b0_0
imageio=2.12.0=pypi_0
importlib-metadata=4.8.1=py38h06a4308_0
importlib-resources=5.4.0=pypi_0
importlib_metadata=4.8.1=hd3eb1b0_0
intel-openmp=2021.4.0=h06a4308_3561
iopath=0.1.9=pypi_0
ipykernel=6.4.1=py38h06a4308_1
ipython=7.29.0=py38hb070fc8_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.5=pyhd3eb1b0_1
isort=4.3.21=pypi_0
jedi=0.18.0=py38h06a4308_1
jinja2=3.0.2=pyhd3eb1b0_0
jpeg=9d=h7f8727e_0
jsonpatch=1.32=pypi_0
jsonpointer=2.2=pypi_0
jsonschema=3.2.0=pyhd3eb1b0_2
jupyter=1.0.0=py38_7
jupyter_client=7.0.6=pyhd3eb1b0_0
jupyter_console=6.4.0=pyhd3eb1b0_0
jupyter_core=4.9.1=py38h06a4308_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
kiwisolver=1.3.2=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libllvm10=10.0.1=hbcb73fb_5
libpng=1.6.37=hbc83047_0
libprotobuf=3.19.1=h4ff587b_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtiff=4.2.0=h85742a9_0
libuuid=1.0.3=h7f8727e_2
libuv=1.40.0=h7b6447c_0
libwebp=1.2.0=h89dd481_0
libwebp-base=1.2.0=h27cfd23_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.12=h03d6c58_0
llvmlite=0.36.0=py38h612dafd_4
lz4-c=1.9.3=h295c915_1
markdown=3.3.4=py38h06a4308_0
markupsafe=2.0.1=py38h27cfd23_0
matplotlib=3.5.0=pypi_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mccabe=0.6.1=pypi_0
mistune=0.8.4=py38h7b6447c_1000
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
multidict=5.2.0=py38h7f8727e_2
mypy-extensions=0.4.3=pypi_0
nbclient=0.5.3=pyhd3eb1b0_0
nbconvert=6.1.0=py38h06a4308_0
nbformat=5.1.3=pyhd3eb1b0_0
ncurses=6.3=h7f8727e_2
nest-asyncio=1.5.1=pyhd3eb1b0_0
networkx=2.6.3=pypi_0
ninja=1.10.2=py38hd09550d_3
notebook=6.4.6=py38h06a4308_0
numba=0.53.1=py38ha9443f7_0
numpy=1.21.2=py38h20f2e39_0
numpy-base=1.21.2=py38h79a1101_0
nvidiacub=1.10.0=0
oauthlib=3.1.1=pyhd3eb1b0_0
olefile=0.46=pyhd3eb1b0_0
omegaconf=2.1.1=pypi_0
opencv-python=4.5.4.60=pypi_0
openssl=1.1.1m=h7f8727e_0
packaging=21.3=pyhd3eb1b0_0
pandocfilters=1.4.3=py38h06a4308_1
parso=0.8.2=pyhd3eb1b0_0
pathspec=0.9.0=pypi_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.4.0=py38h5aabda8_0
pip=21.2.4=py38h06a4308_0
platformdirs=2.4.0=pypi_0
plotly=5.4.0=pypi_0
portalocker=2.3.2=py38h578d9bd_1
prometheus_client=0.12.0=pyhd3eb1b0_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
prompt_toolkit=3.0.20=hd3eb1b0_0
protobuf=3.19.1=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycodestyle=2.8.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pyflakes=2.4.0=pypi_0
pygments=2.10.0=pyhd3eb1b0_0
pyjwt=2.1.0=py38h06a4308_0
pyopenssl=21.0.0=pyhd3eb1b0_1
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.9.2=py38h05f1152_4
pyrsistent=0.18.0=py38heee7806_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.12=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.8=2_cp38
pytorch=1.7.0=py3.8_cuda11.0.221_cudnn8.0.3_0
pytorch3d=0.6.0=pypi_0
pywavelets=1.2.0=pypi_0
pyyaml=5.4.1=py38h497a2fe_0
pyzmq=22.2.1=py38h295c915_1
qt=5.9.7=h5867ecd_1
qtconsole=5.1.1=pyhd3eb1b0_0
qtpy=1.10.0=pyhd3eb1b0_0
readline=8.1=h27cfd23_0
regex=2021.11.10=pypi_0
requests=2.26.0=pypi_0
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
scikit-image=0.18.3=pypi_0
scipy=1.7.2=pypi_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=58.0.4=py38h06a4308_0
setuptools-scm=6.3.2=pypi_0
sip=4.19.13=py38he6710b0_0
six=1.16.0=pyhd3eb1b0_0
sqlite=3.36.0=hc218d9a_0
tabulate=0.8.9=pyhd8ed1ab_0
tbb=2020.3=hfd86e86_0
tenacity=8.0.1=pypi_0
tensorboard=2.6.0=py_1
tensorboard-data-server=0.6.0=py38hca6d32c_0
tensorboard-plugin-wit=1.6.0=py_0
tensorboardx=2.5.1=pypi_0
termcolor=1.1.0=py_2
terminado=0.9.4=py38h06a4308_0
testpath=0.5.0=pyhd3eb1b0_0
tifffile=2021.11.2=pypi_0
tk=8.6.11=h1ccaba5_0
tomli=1.2.2=pypi_0
torchfile=0.1.0=pypi_0
torchvision=0.8.1=py38_cu110
tornado=6.1=py38h27cfd23_0
tqdm=4.62.3=pyhd8ed1ab_0
traitlets=5.1.1=pyhd3eb1b0_0
typing-extensions=3.10.0.2=hd3eb1b0_0
typing_extensions=3.10.0.2=pyh06a4308_0
urllib3=1.26.7=pypi_0
visdom=0.1.8.9=pypi_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py38_1
websocket-client=1.2.1=pypi_0
werkzeug=2.0.2=pyhd3eb1b0_0
wheel=0.37.0=pyhd3eb1b0_1
widgetsnbextension=3.5.1=py38_0
xz=5.2.5=h7b6447c_0
yacs=0.1.8=pypi_0
yaml=0.2.5=h516909a_0
yarl=1.6.3=py38h27cfd23_0
zeromq=4.3.4=h2531618_0
zipp=3.6.0=pyhd3eb1b0_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
